Expose dlv's `--backend` arg in config
https://github.com/derekparker/delve/issues/817

There seems to be an issue using `lldb` as a debugger option. I was able to resolve the issue by changing it to `--backend=native`. To allow more flexibility and unblock atom go-debug users who are experiencing this issue, I suggest that we expose the `--backend` in config while maintaing a default value of `--backend=default`.